scorechainllama/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
streamlit>=1.36.0
jinja2>=3.1.4
requests>=2.32.0
